% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celsiuslids-package.R
\docType{package}
\name{celsiuslids-package}
\alias{celsiuslids}
\alias{celsiuslids-package}
\title{celsiuslids: CeLSIUS Longitudinal Impossible Dataset Generator}
\description{
An R package containing a Shiny application that can be used to generate
'impossible' datasets resembling data extracts from the ONS Longitudinal Study.
}
\section{Main function}{

\code{\link{run_lids_app}} - Launch the LIDS Generator application
}


}
\author{
\strong{Maintainer}: CeLSIUS \email{celsius@ucl.ac.uk}

}
